+++ Example/Emission/index.ios.js
@@ -1,4 +1,4 @@
-import './index.storybooks'
-import '../../src/lib/components/storybooks/'
+import "./index.storybooks"
+import "../../src/lib/Components/Storybooks"
 
-import 'emission';
+import "emission"
