+++ src/lib/Components/Artist/__stories__/index.tsx
@@ -1,2 +1,2 @@
-import "./artist_articles.story"
-import "./artist_header.story"
+import "./ArtistArticles.story"
+import "./ArtistHeader.story"
