+++ src/lib/Components/Inbox/Conversations/__stories__/index.ts
@@ -1,4 +1,4 @@
-import "./artwork_preview.story"
-import "./composer.story.tsx"
-import "./conversation-snippet.story"
-import "./inbox.story"
+import "./ArtworkPreview.story"
+import "./Composer.story.tsx"
+import "./ConversationSnippet.story"
+import "./Inbox.story"
